// frontend/src/openapi/models/ConnectorSyncRun.ts
export interface ConnectorSyncRun {
    /**
     *
     * @type {string}
     * @memberof ConnectorSyncRun
     */
    readonly id: string;
    /**
     *
     * @type {string}
     * @memberof ConnectorSyncRun
     */
    readonly credential: string;
    /**
     *
     * @type {number}
     * @memberof ConnectorSyncRun
     */
    readonly configVersion: number | null;
    /**
     *
     * @type {ConnectorSyncRunModeEnum}
     * @memberof ConnectorSyncRun
     */
    readonly mode: ConnectorSyncRunModeEnum;
    /**
     *
     * @type {Date}
     * @memberof ConnectorSyncRun
     */
    readonly windowFrom: Date | null;
    /**
     *
     * @type {Date}
     * @memberof ConnectorSyncRun
     */
    readonly windowTo: Date | null;
    /**
     *
     * @type {number}
     * @memberof ConnectorSyncRun
     */
    readonly tracesSeen: number;
    /**
     *
     * @type {number}
     * @memberof ConnectorSyncRun
     */
    readonly spansCreated: number;
    /**
     *
     * @type {number}
     * @memberof ConnectorSyncRun
     */
    readonly spansSkipped: number;
    /**
     *
     * @type {ConnectorSyncRunStatusEnum}
     * @memberof ConnectorSyncRun
     */
    readonly status: ConnectorSyncRunStatusEnum;
    /**
     *
     * @type {string}
     * @memberof ConnectorSyncRun
     */
    readonly error: string;
    /**
     *
     * @type {Date}
     * @memberof ConnectorSyncRun
     */
    readonly startedAt: Date;
    /**
     *
     * @type {Date}
     * @memberof ConnectorSyncRun
     */
    readonly finishedAt: Date | null;
}